Refactored exporter module name
